// frontend/src/types/PlayerSearchResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PlayerStatus } from "./PlayerStatus";

export type PlayerSearchResponse = { id: number, server_id: number, server_ip: string, name: string, status: PlayerStatus, last_seen_at: string, };